# apps/docs/public/openapi/v1.0.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sendero Agent Tools",
    "version": "1.0.0",
    "description": "The Sendero tool registry — a single surface for travel booking, treasury rebalancing, OCR, trip-document verification, and guest-escrow settlement.\n\n## Authentication\n\nEvery call is authenticated with a Clerk-issued API key (`Authorization: Bearer ak_…`). Mint one at [`/dashboard/settings/api-keys`](https://www.sendero.travel/dashboard/settings/api-keys). Keys are tenant-scoped; the server derives both `tenantId` and the service-account `userId` from the key.\n\n## Dispatch shape\n\nEvery tool is callable via `POST /api/agent/dispatch` with `{ tool, args, tenantId }` in the body. The same tools are also exposed via MCP at `/api/mcp` for agent runtimes that prefer the MCP protocol.\n\n## Pricing\n\nEach call bills a per-tool nanopayment in micro-USDC via x402. See [`packages/tools/src/pricing.ts`](https://github.com/tcxcx/sendero/blob/main/packages/tools/src/pricing.ts) for the catalog. SaaS plan tiers grant a discount on the nanopayment rate.",
    "contact": {
      "name": "Sendero Developer Experience",
      "url": "https://docs.sendero.travel"
    }
  },
  "servers": [
    {
      "url": "https://www.sendero.travel",
      "description": "Production"
    },
    {
      "url": "https://preview.sendero.travel",
      "description": "Preview"
    }
  ],
  "components": {
    "securitySchemes": {
      "ClerkApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ak_… (Clerk-issued)",
        "description": "Mint a production key at /dashboard/settings/api-keys. Sandbox keys are auto-issued on tenant creation and don't move real USDC."
      }
    },
    "schemas": {
...
      "BookFlightInput": {
        "type": "object",
        "required": ["offerId"],
        "properties": {
          "offerId": {
            "type": "string"
          },
          "services": {
            "type": "array",
            "description": "Ancillary service ids + quantities from list_flight_ancillaries.",
            "items": {
              "type": "object",
              "required": ["id"],
              "properties": {
                "id": {
                  "type": "string"
                },
                "quantity": {
                  "type": "integer",
                  "default": 1,
                  "minimum": 1,
                  "maximum": 9
                }
              }
            }
          },
          "additionalCustomerUserIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Additional icu_… ids to attach (personal assistant, team lead, etc.). The primary traveler is auto-resolved from the session."
          },
          "airlineCreditId": {
            "type": "string",
            "description": "Airline credit id (acd_…) to redeem toward this booking. Remainder paid from balance."
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/agent/dispatch#book_flight": {
      "post": {
        "operationId": "book_flight",
        "summary": "Book a flight for the signed-in user: ensures the traveler identity exists with the supplier, creates a hold order (with any ancillary se…",
        "description": "Book a flight for the signed-in user: ensures the traveler identity exists with the supplier, creates a hold order (with any ancillary services attached), and pays from the pre-funded balance — optionally splitting payment with a traveler airline credit. Returns the PNR + order id. The traveler automatically gets Travel Support Assistant access through their linked identity. Passenger identity comes from the signed-in Clerk session — do not ask the user for name, email, or phone.",
        "tags": ["Bookings"],
        "security": [
          {
            "ClerkApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tenantId", "tool", "args"],
                "properties": {
                  "tenantId": {
                    "type": "string",
                    "description": "The calling tenant's id (tenant_...)."
                  },
                  "userId": {
                    "type": "string",
                    "description": "Optional subject for the call; omitted in service-account mode (we use svc:${keyId} automatically)."
                  },
                  "tool": {
                    "type": "string",
                    "enum": ["book_flight"],
                    "description": "Must be exactly `book_flight` for this operation."
                  },
                  "args": {
                    "$ref": "#/components/schemas/BookFlightInput"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool executed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "result": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "latencyMs": {
                      "type": "number"
                    },
                    "priceMicroUsdc": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "402": {
            "description": "Tenant spend cap exceeded."
          },
          "404": {
            "description": "Tool or tenant not found."
          },
          "500": {
            "description": "Tool execution failed."
          }
        }
      }
    },
...
  }
}